+++ lib/isolator/adapters/background_jobs/active_job.rb
@@ -3,4 +3,8 @@
 Isolator.isolate :active_job,
                  target: ActiveJob::Base,
                  method_name: :enqueue,
-                 exception_class: Isolator::BackgroundJobError
+                 exception_class: Isolator::BackgroundJobError,
+                 details_message: ->(obj, _args) {
+                   "#{obj.class.name}" \
+                   "#{obj.arguments.any? ? " (#{obj.arguments.join(", ")})" : ""}"
+                 }