+++ lib/isolator/adapters/background_jobs/resque.rb
@@ -3,4 +3,7 @@
 Isolator.isolate :resque,
                  target: Resque.singleton_class,
                  method_name: :enqueue,
-                 exception_class: Isolator::BackgroundJobError
+                 exception_class: Isolator::BackgroundJobError,
+                 details_message: ->(_obj, args) {
+                   args.join(", ")
+                 }