+++ lib/isolator/adapters/background_jobs/resque_scheduler.rb
@@ -3,4 +3,7 @@
 Isolator.isolate :resque_scheduler,
                  target: Resque.singleton_class,
                  method_name: :enqueue_at,
-                 exception_class: Isolator::BackgroundJobError
+                 exception_class: Isolator::BackgroundJobError,
+                 details_message: ->(_obj, (ts, *args)) {
+                   "#{args.join(', ')} (at #{ts})"
+                 }