+++ lib/isolator/adapters/background_jobs/sidekiq.rb
@@ -3,4 +3,9 @@
 Isolator.isolate :sidekiq,
                  target: Sidekiq::Client,
                  method_name: :raw_push,
-                 exception_class: Isolator::BackgroundJobError
+                 exception_class: Isolator::BackgroundJobError,
+                 details_message: ->(_obj, args) {
+                   args.first.map do |job|
+                     "#{job['class']} (#{job['args']})"
+                   end.join("\n")
+                 }