+++ lib/isolator/adapters/background_jobs/sucker_punch.rb
@@ -3,4 +3,7 @@
 Isolator.isolate :sucker_punch,
                  target: SuckerPunch::Queue.singleton_class,
                  method_name: :find_or_create,
-                 exception_class: Isolator::BackgroundJobError
+                 exception_class: Isolator::BackgroundJobError,
+                 details_message: ->(_obj, args) {
+                   args.compact.join(", ")
+                 }