+++ lib/isolator/adapters/http/sniffer.rb
@@ -7,7 +7,11 @@ Sniffer::Config.defaults["logger"] = nil
 
 Isolator.isolate :http, target: Sniffer.singleton_class,
                         method_name: :store,
-                        exception_class: Isolator::HTTPError
+                        exception_class: Isolator::HTTPError,
+                        details_message: ->(_obj, args) {
+                          req = args.first.request
+                          "#{req[:method]} #{req[:host]}/#{req[:query]}"
+                        }
 
 Isolator.before_isolate do
   next if Isolator.adapters.http.disabled?