+++ spec/support/workers/active_job_worker.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 class ActiveJobWorker < ActiveJob::Base
-  def perform; end
+  def perform(val = nil); end
 end